completed 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,25 +1,45 @@
 const blogsRouter = require('express').Router()
-const Blog = require('../models/blog')  
+const jwt = require('jsonwebtoken')
+const Blog = require('../models/blog');  
+const User = require('../models/user');
+const config = require('../utils/config');
+const { tokenExtractor, userExtractor } = require('../utils/middleware');
 
+// helper functions
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization');
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '');
+//   }
+//   return null;
+// }
+
+// blogs requests 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({});
+    const blogs = await Blog.find({}).populate('user', "username name id");
     response.json(blogs)
 })
-
   
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body;
 
   if (!body.title || !body.url) {
     response.status(400).end();
   }
 
+  // TODO: get the middleware to take in decodedToken and check for the users
+  const user = request.user;
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user
   });
+
+  user.blogs = user.blogs.concat(blog._id);
+  await user.save();
 
   const savedBlog = await blog.save();
   response.status(201).json(savedBlog)
@@ -39,7 +59,15 @@ blogsRouter.put('/:id', async (request, response) => {
   response.status(200).json(updatedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  
+  // check if logged in 
+  const decodedToken = jwt.verify(request.token, config.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   // if the id is not valid return 400
   const id = request.params.id;
 
@@ -50,6 +78,13 @@ blogsRouter.delete('/:id', async (request, response) => {
   const blog = await Blog.findById(id);
   if (!blog) {
     return response.status(400).end();
+  }
+
+  const user = request.user; 
+
+  // check if the user is the creator of the blog, if not do not allow it
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized' });
   }
 
   await Blog.findByIdAndDelete(id);
